+++ src/pull-scroll/examples/container.vue
@@ -9,7 +9,7 @@
 		<template #header>
 			<div>这是一个容器下的滚动</div>
 		</template>
-		<template #default="it">
+		<template #default="{ it }">
 			<div style="padding: 20px">{{ it }}</div>
 		</template>
 	</vc-pull-scroll>